requirements.txt
python-can
gpsd-py3
